tag delete from the app through contacts

+++ agile-frontend/WebContent/flatfull/jscore/contact-management/contact-details.js
@@ -612,19 +612,28 @@ var Contact_Details_Model_Events = Base_Model_View.extend({
        				
        			// Updates to both model and collection
 	       			App_Contacts.contactDetailView.model.set(data.toJSON(), {silent : true});
-	       			
-	       		//	App_Contacts.contactDetailView.model.set({'tags' : data.get('tags')}, {silent : true}, {merge:false});
-       				
-       				// Also deletes from Tag class if no more contacts are found with this tag
-       				$.ajax({
-       					url: 'core/api/tags/' + tag,
-       					type: 'DELETE',
-       					success: function()
-       					{
-       						if(tagsCollection)
-       							tagsCollection.remove(tagsCollection.where({'tag': tag})[0]);
-       					}
-       				});
+            //  App_Contacts.contactDetailView.model.set({'tags' : data.get('tags')}, {silent : true}, {merge:false});
+              //Check if any deals are having the tag.If yes dont remove it from the app
+              $.ajax({
+                url: 'core/api/opportunity/based/tags?tag=' + tag,
+                type: 'GET',
+                success: function(data)
+                { 
+                  console.log(data);
+                  if(data == "fail"){
+                      // Also deletes from Tag class if no more contacts are found with this tag
+                    $.ajax({
+                      url: 'core/api/tags/' + tag,
+                      type: 'DELETE',
+                      success: function()
+                      {
+                      if(tagsCollection)
+                        tagsCollection.remove(tagsCollection.where({'tag': tag})[0]);
+                      }
+                    });
+                  }
+                }
+              });
        			}
         });
 	
@@ -1192,15 +1201,28 @@ enterCompanyScore: function(e){
 	       		//	App_Contacts.contactDetailView.model.set({'tags' : data.get('tags')}, {silent : true}, {merge:false});
        				
        				// Also deletes from Tag class if no more contacts are found with this tag
-       				$.ajax({
-       					url: 'core/api/tags/' + tag,
-       					type: 'DELETE',
-       					success: function()
-       					{
-       						if(tagsCollection)
-       							tagsCollection.remove(tagsCollection.where({'tag': tag})[0]);
-       					}
-       				});
+
+       			 //Check if any deals are having the tag.If yes dont remove it from the app
+              $.ajax({
+                url: 'core/api/opportunity/based/tags?tag=' + tag,
+                type: 'GET',
+                success: function(data)
+                { 
+                  console.log(data);
+                  if(data == "fail"){
+                      // Also deletes from Tag class if no more contacts are found with this tag
+                    $.ajax({
+                      url: 'core/api/tags/' + tag,
+                      type: 'DELETE',
+                      success: function()
+                      {
+                      if(tagsCollection)
+                        tagsCollection.remove(tagsCollection.where({'tag': tag})[0]);
+                      }
+                    });
+                  }
+                }
+              });
        			}
         });
 	},
